Moved static functions from GraphModel to Util classes
