+++ src/main/java/tudelft/ti2806/pl3/util/DeadEdgeUtil.java
@@ -1,0 +1,44 @@
+package tudelft.ti2806.pl3.util;
+
+import tudelft.ti2806.pl3.data.graph.Edge;
+import tudelft.ti2806.pl3.data.graph.node.Node;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class DeadEdgeUtil {
+	
+	/**
+	 * Removes all edges of which one or both of their nodes is not on the
+	 * graph.
+	 * 
+	 * @param edgeList
+	 *            the list of edges in the graph
+	 * @param nodeList
+	 *            the list of nodes in the graph
+	 */
+	public static void removeAllDeadEdges(List<Edge> edgeList, List<Node> nodeList) {
+		edgeList.removeAll(getAllDeadEdges(edgeList, nodeList));
+	}
+	
+	/**
+	 * Finds all the edges on the graph which have one or two nodes which are
+	 * not on the graph.
+	 * 
+	 * @param edgeList
+	 *            the list of edges in the graph
+	 * @param nodeList
+	 *            the list of nodes in the graph
+	 * @return a list of all dead edges
+	 */
+	public static List<Edge> getAllDeadEdges(List<Edge> edgeList, List<Node> nodeList) {
+		List<Edge> removeList = new ArrayList<Edge>();
+		for (Edge edge : edgeList) {
+			if (!nodeList.contains(edge.getFrom())
+					|| !nodeList.contains(edge.getTo())) {
+				removeList.add(edge);
+			}
+		}
+		return removeList;
+	}
+}
