+++ src/test/java/tudelft/ti2806/pl3/util/DeadEdgeUtilTest.java
@@ -1,4 +1,4 @@
-package tudelft.ti2806.pl3.visualization;
+package tudelft.ti2806.pl3.util;
 
 import org.junit.Assert;
 import org.junit.BeforeClass;
@@ -15,10 +15,8 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-public class GraphModelTest {
+public class DeadEdgeUtilTest {
 	private static SingleNode[] nodes;
-	private static GraphModel dpm;
-	
 	private static List<Node> nodeList;
 	private static List<Edge> edgeList;
 	private static Map<String, Edge> map;
@@ -62,44 +60,6 @@ public class GraphModelTest {
 		edgeList.addAll(map.values());
 		gd = new GraphDataRepository(nodeList, edgeList,
 				new ArrayList<Genome>());
-		dpm = new GraphModel(gd);
-	}
-	
-	@Test
-	public void findFromEdgesTest() {
-		List<Edge> list = dpm.findFromEdges(gd.getEdgeListClone());
-		Assert.assertTrue(list.contains(map.get("1-3")));
-		Assert.assertTrue(list.contains(map.get("2-3")));
-		Assert.assertTrue(list.contains(map.get("3-4")));
-		Assert.assertTrue(list.contains(map.get("4-5")));
-		Assert.assertTrue(list.contains(map.get("7-8")));
-		Assert.assertTrue(list.contains(map.get("8-9")));
-		Assert.assertTrue(list.size() == 6);
-	}
-	
-	@Test
-	public void findToEdgesTest() {
-		List<Edge> list = dpm.findToEdges(gd.getEdgeListClone());
-		Assert.assertTrue(list.contains(map.get("0-1")));
-		Assert.assertTrue(list.contains(map.get("0-2")));
-		Assert.assertTrue(list.contains(map.get("3-4")));
-		Assert.assertTrue(list.contains(map.get("4-5")));
-		Assert.assertTrue(list.contains(map.get("5-6")));
-		Assert.assertTrue(list.contains(map.get("5-7")));
-		Assert.assertTrue(list.contains(map.get("7-8")));
-		Assert.assertTrue(list.contains(map.get("8-9")));
-		Assert.assertTrue(list.size() == 8);
-	}
-	
-	@Test
-	public void findCombinableNodesTest() {
-		List<Edge> list = dpm.findCombineableNodes(gd.getNodeListClone(),
-				gd.getEdgeListClone());
-		Assert.assertTrue(list.contains(map.get("3-4")));
-		Assert.assertTrue(list.contains(map.get("4-5")));
-		Assert.assertTrue(list.contains(map.get("7-8")));
-		Assert.assertTrue(list.contains(map.get("8-9")));
-		Assert.assertTrue(list.size() == 4);
 	}
 	
 	@Test
@@ -107,17 +67,7 @@ public class GraphModelTest {
 		Edge deadEdge = new Edge(nodes[0], new SingleNode(-1, null, 0, 0, null));
 		List<Edge> edgeList = gd.getEdgeListClone();
 		edgeList.add(deadEdge);
-		GraphModel.removeAllDeadEdges(edgeList, gd.getNodeListClone());
+		DeadEdgeUtil.removeAllDeadEdges(edgeList, gd.getNodeListClone());
 		Assert.assertFalse(edgeList.contains(deadEdge));
-	}
-	
-	@Test
-	public void combineNodesTest() {
-		List<Node> nodeList = gd.getNodeListClone();
-		List<Edge> edgeList = gd.getEdgeListClone();
-		dpm.combineNodes(dpm.findCombineableNodes(nodeList, edgeList),
-				nodeList, edgeList);
-		Assert.assertTrue(nodeList.size() == 6);
-		Assert.assertTrue(edgeList.size() == 6);
 	}
 }
